Refactored projects container
