+++ src/features/projects/screens/Projects.tsx
@@ -1,6 +1,12 @@
 import React, { ReactElement } from 'react';
-import ProjectsContainer from '../components/ProjectsContainer';
 import styles from '../styles/Projects.module.scss';
+import dynamic from 'next/dynamic';
+import ProjectLoader from '../../common/ContentLoaders/Projects/ProjectLoader';
+import i18next from '../../../../i18n/'
+import LazyLoad from 'react-lazyload';
+import NotFound from '../../../../public/assets/images/NotFound';
+import Header from '../components/projects/Header';
+import SearchBar from '../components/projects/SearchBar';
 
 interface Props {
   projects: any;
@@ -11,6 +17,11 @@ interface Props {
   setsearchedProjects: any
 }
 
+const { useTranslation } = i18next;
+const ProjectSnippet = dynamic(() => import('../components/ProjectSnippet'), {
+  loading: () => <ProjectLoader />,
+});
+
 function ProjectsList({
   projects,
   directGift,
@@ -19,20 +30,115 @@ function ProjectsList({
   setShowProjects,
   setsearchedProjects,
 }: Props): ReactElement {
-  const [searchedProjects, setSearchedProjects] = React.useState([]);
-  const [allProjects, setAllProjects] = React.useState(projects);
+  
   const screenWidth = window.innerWidth;
   const screenHeight = window.innerHeight;
   const isMobile = screenWidth <= 767;
   const [scrollY, setScrollY] = React.useState(0);
+  const { t } = useTranslation(['donate']);
 
-  const ProjectsProps = {
+
+  const featuredList = process.env.NEXT_PUBLIC_FEATURED_LIST;
+
+  const showFeaturedList =
+    featuredList === 'false' || featuredList === '0' ? false : true;
+
+  const [selectedTab, setSelectedTab] = React.useState('all');
+  const [searchMode, setSearchMode] = React.useState(false);
+
+  React.useEffect(() => {
+    showFeaturedList ? setSelectedTab('featured') : null;
+  }, []);
+
+  const [searchValue, setSearchValue] = React.useState('');
+
+  const searchRef = React.useRef(null);
+
+  function getProjects(projects: Array<any>, type: string) {
+    if (type === 'featured') {
+      return projects.filter(
+        (project: { properties: { isFeatured: boolean } }) =>
+          project.properties.isFeatured === true
+      );
+    } else if (type === 'all') {
+      return projects;
+    }
+  }
+
+  function getSearchProjects(projects: Array<any>, keyword: string) {
+    let resultProjects = [];
+    if (keyword !== '') {
+      resultProjects = projects.filter(function (project) {
+        if (
+          project.properties.name.toLowerCase().includes(keyword.toLowerCase())
+        ) {
+          return true;
+        } else if (
+          project.properties.location &&
+          project.properties.location
+            .toLowerCase()
+            .includes(keyword.toLowerCase())
+        ) {
+          return true;
+        } else if (
+          project.properties.tpo.name &&
+          project.properties.tpo.name
+            .toLowerCase()
+            .includes(keyword.toLowerCase())
+        ) {
+          return true;
+        } else {
+          return false;
+        }
+      });
+      setsearchedProjects(resultProjects);
+      return resultProjects;
+    } else {
+      setsearchedProjects(projects);
+    }
+  }
+
+  const allProjects = React.useMemo(() => getProjects(projects, 'all'), [
     projects,
-    setSearchedProjects: setsearchedProjects,
-    directGift,
-    setDirectGift,
-    setShowProjects,
-  };
+  ]);
+
+  const searchProjectResults = React.useMemo(
+    () => getSearchProjects(projects, searchValue),
+    [searchValue]
+  );
+
+  const featuredProjects = React.useMemo(
+    () => getProjects(projects, 'featured'),
+    [projects]
+  );
+
+  const AllProjects = (projects:any)=>{   
+    if (projects.projects.length < 1) {
+      return (
+        <div className={styles.projectNotFound}>
+          <LazyLoad>
+            <NotFound className={styles.projectNotFoundImage} />
+            <h5>{t('donate:noProjectsFound')}</h5>
+          </LazyLoad>
+        </div>
+      );
+    }
+    else {
+      return (
+        projects.projects.map((project: any) => {
+          return (
+            <ProjectSnippet
+              key={project.properties.id}
+              project={project}
+              directGift={directGift}
+              setDirectGift={setDirectGift}
+              editMode={false}
+            />
+          );
+        })
+      ) 
+    } 
+  }
 
   return (
     <>
@@ -50,7 +156,33 @@ function ProjectsList({
             }
           }}
         >
-          <ProjectsContainer {...ProjectsProps} />
+          <div className={styles.header} style={isMobile ? { height: '66px', paddingTop: '16px' } : {}}>
+            {isMobile ? <div className={styles.dragBar}></div> : null}
+            {searchMode ? 
+              <SearchBar 
+                setSearchValue={setSearchValue} 
+                setSearchMode={setSearchMode} 
+                searchValue={searchValue} 
+                searchRef={searchRef}
+              /> :  
+              <Header 
+                showFeaturedList={showFeaturedList} 
+                setSelectedTab={setSelectedTab} 
+                selectedTab={selectedTab} 
+                setSearchMode={setSearchMode} 
+                projects={projects} 
+              />}
+            </div>
+            {/* till here is header */}
+            <div className={styles.projectsContainer}>
+              <div className={styles.projectsContainerChild}>
+              {searchValue !== '' ?
+                <AllProjects projects={searchProjectResults} />
+                : selectedTab === 'all' ? 
+                  <AllProjects projects={allProjects} /> : 
+                  <AllProjects projects={featuredProjects} />}
+              </div>        
+            </div>
         </div>
       ) : null}
     </>
